Defensively check whether we found a table during spec creation
It seems possible in some cases to have a table driven scenario that fails in a way that means we don't have a ProtoItem_Table in the items.

It seems there is a probably an issue upstream in Gauge somehow, but this seems low risk to check defensively, even if it leads to an incomplete report?

Signed-off-by: Chad Wilson <29788154+chadlwilson@users.noreply.github.com>